group_special change
I *think* this may fix the group_special issue #320.
I don't have a way to test this out right now, but I think this may fix the problem since main.py code uses the ArgParser to alias special-card-groups to group_special.  In this case ArgParser is being bypassed, I think.  So the variable to set is group_special.